Extract dashboard UI, add AI payment risk panel
